requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
youtube-transcript-api
chromadb